Big patch with header file restructuring, cleanups, and some added comments.
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1479 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/resource/getres.c
@@ -289,7 +289,8 @@ load_package (INDEX_HEADERPTR ResHeaderPtr, RES_PACKAGE res_package)
 #ifdef DEBUG
 					if (hData == 0)
 						fprintf (stderr, "fp = 0x%08lx(%d), type = %u, "
-								"file_offs = %lu, len = %lu\n", fp,
+								"file_offs = %lu, len = %lu\n",
+								(unsigned long) fp,
 								DanglingOpen, type, file_offs, data_len);
 #endif /* DEBUG */
 				}
